refactor(dispatch/view): rearchitecture handling with views; move views out of their folders
