--- telefone/framework/dispatch/view/abc/__init__.py
@@ -1,6 +1,0 @@
-from .dispense import ABCDispenseView
-from .message import ABCMessageView
-from .raw import ABCRawUpdateView
-from .view import ABCView
-
-__all__ = ("ABCDispenseView", "ABCMessageView", "ABCRawUpdateView", "ABCView")
