--- telefone/framework/dispatch/view/default/__init__.py
@@ -1,9 +1,0 @@
-from .message import ABCBotMessageView, BotMessageView
-from .raw import BotHandlerBasement, BotRawUpdateView
-
-__all__ = (
-    "ABCBotMessageView",
-    "BotMessageView",
-    "BotHandlerBasement",
-    "BotRawUpdateView",
-)
